added if then
if then is based on user savedsetting being yes or no. if saved setting
is yes then nothing changes and a console message is written. if saved
setting is no then it erases everything by rewriting the div tags
surrounding the whole doc to just have a message to download
automatically.

We still need to implement the auto download for when saved setting is
no and when saved setting is yes and we need the submit button to work.